# Performance fixes