--- a/src/main/java/shortestpath/pathfinder/CollisionMap.java
+++ b/src/main/java/shortestpath/pathfinder/CollisionMap.java
@@ -2,18 +2,26 @@ package shortestpath.pathfinder;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import net.runelite.api.coords.WorldPoint;
 import shortestpath.ShortestPathPlugin;
 import shortestpath.Transport;
 import shortestpath.Util;
 
 public class CollisionMap extends SplitFlagMap {
+
     // Enum.values() makes copies every time which hurts performance in the hotpath
     private static final OrdinalDirection[] ORDINAL_VALUES = OrdinalDirection.values();
+    private static RegionExtent regionExtents = new RegionExtent(0, 0, 0, 0);
 
     public CollisionMap(Map<Integer, byte[]> compressedRegions) {
         super(compressedRegions, 2);
@@ -68,7 +76,7 @@ public class CollisionMap extends SplitFlagMap {
 
         // Transports are pre-filtered by PathfinderConfig.refreshTransportData
         // Thus any transports in the list are guaranteed to be valid per the user's settings
-        for (Transport transport : config.getTransports().getOrDefault(node.position, (List<Transport>)Collections.EMPTY_LIST)) {
+        for (Transport transport : config.getTransports().getOrDefault(node.position, (List<Transport>) Collections.EMPTY_LIST)) {
             neighbors.add(new TransportNode(transport.getDestination(), node, transport.getWait()));
         }
 
@@ -116,18 +124,49 @@ public class CollisionMap extends SplitFlagMap {
     public static CollisionMap fromResources() {
         Map<Integer, byte[]> compressedRegions = new HashMap<>();
         try (ZipInputStream in = new ZipInputStream(ShortestPathPlugin.class.getResourceAsStream("/collision-map.zip"))) {
+            int minX = Integer.MAX_VALUE;
+            int minY = Integer.MAX_VALUE;
+            int maxX = 0;
+            int maxY = 0;
+
             ZipEntry entry;
             while ((entry = in.getNextEntry()) != null) {
                 String[] n = entry.getName().split("_");
+                final int x = Integer.parseInt(n[0]);
+                final int y = Integer.parseInt(n[1]);
+                minX = Math.min(minX, x);
+                minY = Math.min(minY, y);
+                maxX = Math.max(maxX, x);
+                maxY = Math.max(maxY, y);
 
                 compressedRegions.put(
-                        SplitFlagMap.packPosition(Integer.parseInt(n[0]), Integer.parseInt(n[1])),
+                        SplitFlagMap.packPosition(x, y),
                         Util.readAllBytes(in)
                 );
             }
+
+            regionExtents = new RegionExtent(minX, minY, maxX, maxY);
         } catch (IOException e) {
             throw new UncheckedIOException(e);
         }
         return new CollisionMap(compressedRegions);
+    }
+
+    public static RegionExtent getRegionExtents() {
+        return regionExtents;
+    }
+
+    @RequiredArgsConstructor
+    @Getter
+    public static class RegionExtent {
+        public final int minX, minY, maxX, maxY;
+
+        public int getWidth() {
+            return maxX - minX;
+        }
+
+        public int getHeight() {
+            return maxY - minY;
+        }
     }
 }