--- a/src/main/java/shortestpath/pathfinder/Pathfinder.java
+++ b/src/main/java/shortestpath/pathfinder/Pathfinder.java
@@ -1,14 +1,8 @@
 package shortestpath.pathfinder;
 
 import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Deque;
-import java.util.HashSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.PriorityQueue;
-import java.util.Queue;
-import java.util.Set;
+import java.util.*;
+
 import lombok.Getter;
 import net.runelite.api.coords.WorldPoint;
 
@@ -18,10 +12,13 @@ public class Pathfinder implements Runnable {
     @Getter
     private final WorldPoint target;
     private final PathfinderConfig config;
+    private final boolean targetInWilderness;
 
-    private final Deque<Node> boundary = new LinkedList<>();
-    private final Set<WorldPoint> visited = new HashSet<>();
-    private final Queue<Node> pending = new PriorityQueue<>();
+    // Capacities should be enough to store all nodes without requiring the queue to grow
+    // They were found by checking the max queue size
+    private final Deque<Node> boundary = new ArrayDeque<>(4096);
+    private final Queue<Node> pending = new PriorityQueue<>(256);
+    private final VisitedTiles visited = new VisitedTiles();
 
     @Getter
     private List<WorldPoint> path = new ArrayList<>();
@@ -33,16 +30,17 @@ public class Pathfinder implements Runnable {
         this.start = start;
         this.target = target;
         this.config.refresh();
+        targetInWilderness = PathfinderConfig.isInWilderness(target);
 
         new Thread(this).start();
     }
 
     private void addNeighbors(Node node) {
         for (Node neighbor : config.getMap().getNeighbors(node, config)) {
-            if (config.avoidWilderness(node.position, neighbor.position, target)) {
+            if (!visited.get(node.position) && config.avoidWilderness(node.position, neighbor.position, targetInWilderness)) {
                 continue;
             }
-            if (visited.add(neighbor.position)) {
+            if (visited.set(neighbor.position)) {
                 if (neighbor instanceof TransportNode) {
                     pending.add(neighbor);
                 } else {