--- a/src/main/java/shortestpath/pathfinder/PathfinderConfig.java
+++ b/src/main/java/shortestpath/pathfinder/PathfinderConfig.java
@@ -33,6 +33,7 @@ public class PathfinderConfig {
 
     @Getter
     private Duration calculationCutoff;
+    @Getter
     private boolean avoidWilderness;
     private boolean useAgilityShortcuts;
     private boolean useGrappleShortcuts;