Optimized pathfinding as much as I could without making any major refactors:

- Preloads all transports at startup and filters transports based on config settings before pathfinding rather than during. This saves a bit of time for each search and makes adding more transports a bit less expensive.
  - Transport options still function the same; `PathfinderConfig.useTransport` already checked if it should use various modes of transport so the checks in `Transport.fromResources` were redundant.
- Optimize data structures:
  - Using a `HashSet` for tracking visited nodes was expensive taking up 50% of the total execution time. I've replaced this with `VisitedTiles` which behaves the same but with significantly less overhead (and fewer allocations :) ).
  - Replaced `LinkedList` with `ArrayDeque`. `ArrayDeque` uses a circular array so inserting elements to the front or back of the queue is free. I've set conservative initial capacities for both queues that should avoid the need to grow either queue during the search.
- Remove `SplitFlagMap.Position` and used a packed Integer instead. Doesn't help performance all too much but significantly cuts back on the amount of data being allocated per search.
  - Using a sparse array instead of a `HashMap` for the `compressedRegion` map would likely speed up lookups but I don't think it's necessary as the cache lookups are the main performance cost.
- Cache `OrdinalDirection.values()` as every call makes an array copy despite the values never changing. This hurt both runtime performance and made many unnecessary allocations.
- `PathfinderConfig.isInWilderness()` is moderately expensive and makes many allocations so I avoid calculating it or cache the target wilderness status when possible.

# Results
Ran through a bunch of different paths short and long in-game to make sure the resulting path is still the same as OSRS's pathfinding.

**Worst-Case** test: from GE to top-left corner of the map in the middle of the ocean
**Short-Case** test: from GE to a point just west 105 tiles away

Ran a few paths before-hand to warm-up the JVM before recording results.
Screenshots are of the Intellij profiler of both CPU Samples and Memory Allocations

System used is an Intel i9-13900K with 64GB DDR5 RAM, OpenJDK 11.0.12.0 (a bit out of date). While not the average system by far, results should scale similarly on other machines.

| Metric | Before (1740a91) | After (6137611) | Difference |
| --- | --- | --- | --- |
**Worst-Case Time** | 2200.5582ms | 1224.3621ms | **976.1961ms**
**Short-Case Time** | 44.2161ms | 28.5745ms | **15.6416ms**
**Allocations** | 3,741,112,192 bytes | 1,314,359,520 bytes | **2,426,752,672 bytes**

## Before Optimizations

**Worst-Case**: 2200.5582ms
**Short-Case**: 44.2161ms

**CPU Samples (Worst-Case):**
<img width="465" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/fb1995dd-619d-4bad-abed-30fee252500b">

**Memory Allocations (Worst-Case):**
<img width="628" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/f18bec65-9e0e-4a7d-9a64-4b4378dbe50f">

## After Optimizations

**Worst-Case**: 1224.3621ms
**Short-Case**: 28.5745ms

**CPU Samples (Worst-Case):**
<img width="392" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/e2b1eb1d-f43e-4332-b056-19dc4c02fd1f">

**Memory Allocations (Worst-Case):**
<img width="613" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/26e4b109-33f3-48ad-b189-10f735d3ae56">