@jocopa3 2023-07-15
Addressed comments and cleaned up a few minor things.