@Skretzo 2023-07-15
I can't seem to spot exactly where, but when testing this PR the `Avoid wilderness` option in the config no longer makes any difference for the pathfinding. When this option is enabled the path should not use the wilderness lever to go from Edgeville to Ardougne.
![pathfinding_avoid_wilderness_config_not_working](https://github.com/Skretzo/shortest-path/assets/53493631/6986d917-0f04-411d-bbb8-bce7180a045d)