@jocopa3 2023-07-15
> the Avoid wilderness option in the config no longer makes any difference for the pathfinding

I've found and fixed this though there are some weird results like below; I'm not sure if this is expected or not but it seems like a quirk of the closest-point heuristic:
<img width="461" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/e8b762bd-9f9b-4550-a35c-3fa7919c3394">

Performance wise in the worst-case, runtime speed is the same with/without due to the reduction in search space while allocations rose to 1,402,018,656 bytes from 1,314,359,520 bytes. Allocation impact can be eliminated with the `WorldPoint` packing optimization I've mentioned but I'm not going to do that in this PR.

<img width="692" alt="image" src="https://github.com/Skretzo/shortest-path/assets/2230152/a39d0053-bcc2-4724-965b-c0b803d05cfd">