@Skretzo 2023-07-15
> I'm not sure if this is expected or not but it seems like a quirk of the closest-point heuristic

Believe it or not, but the path you got right there is more or less correct. You probably had `Avoid wilderness` enabled and `Use fairy rings` disabled. I have not added all agility shortcuts and teleports in the game yet, so the closest dungeon to the location you picked out in the ocean (cyan mark) was therefore Tarn's Lair (green circle), which requires a bit of a winding path through Morytania and the Haunted Mine when starting at the GE (blue circle).

![image](https://github.com/Skretzo/shortest-path/assets/53493631/dfbf23d2-9dd5-4f73-9765-36126aab39e6)

If you want a better map than the world map in-game I would highly recommend https://mejrs.github.io/osrs. That map allows you to see all the dungeons up north of the main world, as well as displaying coordinates, labels, crowdsourced transport locations, objects, nomove tiles and npc information.